Created app for custom users and moved Profile model to it
