+++ ledger/models.py
@@ -1,15 +1,6 @@
-from django.contrib.auth.models import User
 from django.db import models
 from django.urls import reverse
-
-
-class Profile(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    name = models.CharField(max_length=50)
-    bio = models.CharField(max_length=255)
-
-    def __str__(self):
-        return self.name
+from user.models import Profile
 
 
 class Ingredient(models.Model):
